feat(Product) : Jpa 연동
- InMemoryProductRepository @Component 제거, Service 테스트를 위해 남겨둠.
- ProductJpaRepository, ProductRepositoryImpl 생성
- ProductEntity : update 메서드 추가
- ProductRepositoryTest : InMemory에서 @DataJpaTest로 변경
- V1__init.sql : product.price int -> DECIMAl로 변경.

+++ src/main/resources/db/migration/V1__init.sql
@@ -13,7 +13,7 @@ CREATE TABLE product
 (
     id         BIGINT         NOT NULL AUTO_INCREMENT,
     name       VARCHAR(255)   NOT NULL,
-    price      INT            NOT NULL,
+    price      DECIMAL(13, 2) NOT NULL,
     image_url  VARCHAR(2048)  NOT NULL,
     created_at DATETIME(6)    NOT NULL DEFAULT CURRENT_TIMESTAMP(6),
     updated_at DATETIME(6)    NOT NULL DEFAULT CURRENT_TIMESTAMP(6) ON UPDATE CURRENT_TIMESTAMP(6),
